+++ Changes.md
@@ -9,6 +9,7 @@ Improvements
 Fixes
 -----
 
+- Reference : Fixed loss of additional CompoundDataPlug children during reloading.
 - UIEditor : Fixed Python 3 compatibility in plug presets editor.
 - NodeGadget : Fixed intermittent shutdown crash.
 - Expression : Fixed handling of expressions which assign CompoundData to `Attributes.extraAttributes`.